\name{Autotune}
\alias{Autotune}
\alias{Autotune}
\docType{package}
\title{
Autotune LASSO
}
\description{
Autotuning LASSO
}
\details{
\tabular{ll}{
Package: \tab AutotuneRpackage\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2024-12-13\cr
License: \tab What license is it under? \cr
}
~~ There are 3 main functions in this package
autotune_lasso: Main Autotune LASSO
autotune_lasso_active: Autotune LASSO with active set selection
autotune_lasso_l2: Autotune LASSO which works with the l2 norm of partial residuals
~~
}
\author{
Tathagata Sadhukhan, Ines Wilms, Stephan Smeekes, Sumanta Basu
Maintainer: Tathagata Sadhukhan <ts767@cornell.edu>
}
