% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotune_lasso.R
\name{autotune_Lasso}
\alias{autotune_Lasso}
\title{Autotune LASSO: Fitting a linear model with fast and automatic lasso regularization}
\arguments{
\item{xin}{Matrix of predictors, with one column for each predictor.
Dimension will be \code{nobs} \eqn{\times} \code{nvars}; so each row is a new observation
and \code{nvars} >1.}

\item{yin}{Vector of responses.}

\item{alpha}{Default 0.01, significance level of sequential F-tests 
used for estimation of support set.}

\item{lambda0}{Numeric input, default (and recommended) value None.
If user provides a \eqn{\lambda}, autotune lasso starts finding pathwise solutions
from that \eqn{\lambda}, else sets the starting \eqn{\lambda = \|x^T y\|_{\infty}/}(\code{nobs} \times \eqn{Var(y)})}

\item{tolerance}{Numeric input for an additional stopping criteria on the
coordinate descent when sigma is updating. Stops that coordinate
descent when the relative change between successive iterates of
coefficients' estimates is less than the \code{tolerance}.}

\item{beta_tolerance}{Numeric input for the stopping criteria on the
coordinate descent when sigma is not updating. Stops that coordinate
descent when the relative change between successive iterates of
coefficients' estimates is less than the \code{beta_tolerance}.}

\item{iter_max}{Maximum number of iterations of coordinate descent
allowed when sigma is updating}

\item{beta_iter_max}{Maximum number of iterations of coordinate 
descent allowed when sigma is not updating}

\item{standardize}{Logical flag for standardization of all variables in x, prior to
fitting the model sequence. The coefficients are always returned on the
original scale.}

\item{standardize_response}{Logical flag for demeaning the reponse variable y.}

\item{intercept}{Should intercept(s) be fitted (default=TRUE) or set to zero
(FALSE).}

\item{active}{Should active set selection be used (TRUE) or avoided
(default = FALSE). Autotune fast convergence often alleviates the speed gains
coming from active set selection.}

\item{verbose}{Logical input, default \code{FALSE}; if \code{TRUE}
prints out the following details of Autotune LASSO path:
number of iterations of coordinate descent when sigma was updating,
number of iterations of coordinate descent when sigma was not updating,
and the size of the estimated support set(or the number of outliers
identified in the final distribution of \eqn{\ell_1} norm of partial residuals).}

\item{l2_partial_residuals}{Logical flag to whether use the l2 norm of partial 
residuals for ordering them instead of the default l1 norm.}
}
\value{
A list with various intermediate and final outputs produced by Autotune LASSO in its regularization path.
\item{beta_mat}{A (\code{no_of_iter} + 1) \eqn{ \times} \code{nvars} matrix
consisting of coefficients after the end of each iteration of
coordinate descent while sigma was updating and the last column
have the final estimates of coefficients after coordinate descent has converged.} 
\item{sigma2_seq}{ A \code{no_of_iter} length sequence of estimates
noise variance \eqn{\sigma^2}.}
\item{beta}{ Final estimates of regression coefficients. }
\item{lambda0}{ Value of \eqn{\lambda_0} used in the Autotune LASSO.
Refer to the original paper for details.}
\item{lambda}{Final thresholding value \eqn{\lambda =
\lambda_0\hat\sigma^2} used in the coordinate descent after noise
variance estimate \eqn{\hat\sigma^2} has converged.}
\item{sorted_predictors}{Decreasing ordering of predictors in terms of
their contribution to predicting the response values.}
\item{support_set}{Final set of predictors included in the support set by autotune lasso.}
\item{no_of_iter}{Number of iterations of coordinate descent performe
before noise variance estimate \eqn{\hat{\sigma}^2} converged.}
\item{no_of_beta_iter}{After noise variance estimate \eqn{\hat{\sigma}^2} has
converged, it is the number of iterations of coordinate descent required
for coefficients \eqn{\hat\beta} to converge.}
\item{count_sig_beta}{(\code{no_of_iter})-length vector containing the
support set sizes across the coordinate descent iterations while the
noise variance estimate is being updated.}
}
\description{
Fits a linear model via alternative optimization of penalized gaussian maximum likelihood 
which is a biconvex function of regression coefficients \eqn{\beta} and noise variance \eqn{\sigma^2}. 
Autotune LASSO's regularization path quickly picks out a good lambda for LASSO and then
returns the corresponding linear fit along with various attributes related to the fit.
}
\examples{
set.seed(10)
n <- 80
p <- 500
s <- 5
type <- 1
snr <- 3
betatrue <- c(rep(1,s), rep(0, p - s))
x <- scale(matrix(rnorm(n * p), ncol = p))
error.sd <- sqrt((betatrue \%*\% betatrue)/snr)
err <- rnorm(n, sd = error.sd)
y <- x \%*\% betatrue + err
y <- y - mean(y)
ans <- autotune_lasso(x, y, verbose = TRUE)
b <- betatrue
# The Predictors which are actually significant:
which(b != 0)
# The Predictors which had nonzero estmated coefficients:
which(ans$beta != 0)
# Top 10 predictors X_i's in the ranking of X_i's given by autotune:
ans$sorted_predictors[1:10] + 1
# No of significant predictors in each CD iteration when sigma_hat is allowed to vary:
ans$count_sig_beta
# Sigma estimates in each CD iteration:
ans$sigma2_seq
# Empirical noise variance:
var(err)


}
